Clean up sriForVersion util
