+++ src/routes/library.ts
@@ -11,14 +11,52 @@ import {
 } from '../utils/kvMetadata.ts';
 import { queryCheck } from '../utils/query.ts';
 import respond, { notFound } from '../utils/respond.ts';
-import sriForVersion from '../utils/sriForVersion.ts';
 
 import type {
     LibraryResponse,
     LibraryVersionResponse,
 } from './library.schema.ts';
 
-const extensions = Object.keys(files);
+/**
+ * Create a map of file names to SRI hashes, based on library files and SRI data.
+ *
+ * @param library Name of the library.
+ * @param version Version of the library.
+ * @param files Names of the files for this version of the library.
+ * @param sriData SRI data for the library version.
+ */
+const sriForVersion = (
+    library: string,
+    version: string,
+    files: string[],
+    sriData: Record<string, string>,
+) => {
+    // Build the SRI object
+    const sri: Record<string, string> = {};
+    for (const file of files) {
+        const fullFile = `${library}/${version}/${file}`;
+
+        // If we have an SRI entry for this, add it
+        if (sriData && sriData[fullFile]) {
+            sri[file] = sriData[fullFile];
+            continue;
+        }
+
+        // If we don't have an SRI entry, but expect one, error!
+        // if (file.endsWith('.js') || file.endsWith('.css')) {
+        //     Sentry.withScope(scope => {
+        //         scope.setTag('library', library);
+        //         scope.setTag('library.version', version);
+        //         scope.setTag('library.file', file);
+        //         scope.setTag('library.file.full', fullFile);
+        //         Sentry.captureException(new Error('Missing SRI entry'));
+        //     });
+        // }
+    }
+
+    // Done
+    return sri;
+};
 
 /**
  * Check if a file is whitelisted for cdnjs, based on its extension.
@@ -26,7 +64,7 @@ const extensions = Object.keys(files);
  * @param file Filename to check.
  */
 const whitelisted = (file: string) =>
-    extensions.includes(file.split('.').slice(-1)[0] || '');
+    Object.keys(files).includes(file.split('.').slice(-1)[0] || '');
 
 /**
  * Handle GET /libraries/:library/:version requests.
